Enable adding new categories, new commands. Add sidebar nav and remove the default navbar

+++ Brain.Api/Controllers/CommandsController.cs
@@ -1,30 +1,44 @@
 ﻿using System;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Brain.Api.Models;
+using Brain.Api.Models.Account;
 using Brain.Api.Repositories;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Brain.Api.Controllers
 {
     [ApiController]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     [Route("/api/commands")]
     public class CommandsController : ControllerBase
     {
         private readonly CommandsRepository _commands;
+        
+        private readonly UserManager<User> _userManager;
 
-        public CommandsController(CommandsRepository commandsRepository) 
+        public CommandsController(CommandsRepository commandsRepository, UserManager<User> userManager) 
         {
             _commands = commandsRepository;
+            
+            _userManager = userManager;
         }
-
-        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+        
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
-            var commands = await _commands.GetAll();
-            return Ok(commands);
+            var jwtUser = HttpContext.User;
+            var userId= jwtUser.FindFirst(ClaimTypes.NameIdentifier).Value;
+            if (userId != null)
+            {
+                var user = await _userManager.FindByIdAsync(userId);
+                var commands = await _commands.GetAll(user);
+                return Ok(commands);
+            }
+            return Unauthorized();
         }
 
         [HttpGet("{id:int}")]
@@ -53,14 +67,20 @@ namespace Brain.Api.Controllers
         {
             try
             {
-                var command = await _commands.DeleteAsync(id);
-                return NoContent();
+                var jwtUser = HttpContext.User;
+                var userId = jwtUser.FindFirst(ClaimTypes.NameIdentifier).Value;
+                if (userId != null)
+                {
+                    await _commands.DeleteAsync(id);
+                    return NoContent();
+                }
             }
             catch (Exception)
             {
                 return NotFound();
             }
 
+            return BadRequest();
         }
 
         [HttpPatch]
